backend/pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = "The backend for the new IITH Dashboard"
authors = ["Lambda <lambda@iith.ac.in>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
google-auth = "^2.26.1"
pypika = "^0.48.9"
requests = "^2.31.0"
pyjwt = "^2.8.0"
gspread = "^6.1.2"
numpy = "^2.0.0"
oauth2client = "^4.1.3"
python-multipart = "^0.0.9"
boto3 = "^1.34.131"
elasticsearch = "^8.14.0"
html2text = "^2024.2.26"
aiosql = "^10.2"
gunicorn = "^22.0.0"
pytz = "^2024.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
